third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-audio-element/audio-loading-autoplay-deferred.tentative-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_unreached: Audio should not fire loadstart event when not visible in viewport Reached unreachable code
[FAIL] Audio with loading=lazy and autoplay that is not visible in viewport does not load audio data or autoplay
  assert_equals: expected 0 but got 4
Harness: the test ran to completion.
